Ensure loading of local translation bundles doesn't override a cached bundle with higher version
It's not necessary, but may be a good idea to install into cache, the latest local translation bundle. This way we can fall back onto something (say we packaged a functional copy, but later downloaded a broken one...). However, upon installing - only select it as current if there is no higher cached version.