TML should be questioned for whether inline translations are enabled
Asking application may be fruitless when there is no bundle, and therefore application object, available. However, we may be able to communicate with the API server with respect to inline translations; and when that happens via TML.enableTranslations - this state may be different. It's a bit awkward at the moment...