Let's not store default language in a separate property
that's asking for data disparity with defaultLocale and possibly languages, if not being careful. Instead, we resort to keeping default locale (from API) and array of languages (from API, as TMLLanguage objects) and then matching those languages's locale with default locale. We'll cache the result in an ivar, to avoid unnecessary lookups, as ppl tend to love to wear our their keyboard keys 's', 'e', 'l', 'f', '.', causing unnecessary lookups...