Add ability to enabled/disable TMLAPIBundle's sync capability
The reason behind it is that - when an app starts, localization takes place, and when translations aren't found - missing translation keys need to be recorded. We can't ignore this as we did based on the criteria that we're not using the API bundle. We should record missing strings, and if we switch to API bundle - missing translations will get sorted out...

In that sense, we'll always add missing translation keys to the API bundle, but it won't sync as the result of adding missing translations, until we make the bundle current...