Using long press gesture to translate things within the app
Instead of adding a million gesture recognizers as it used to be - we add just one - to the key window. When the long press gesture recognizer begins (we don't want to wait for fingers to lift the screen - that's just weird) - we'll look for a subview that has string to localize. If we find at least one string (and currently we only handle one...) - present translator view controller.

Change some of the weirdness in the translator view controller's view makeup. We're wrapping that controller in the navigation controller, and setting appropriate navigation items in the translator view controller, as expected - instead of mocking objects and throwing them up into the view hierarchy - which doesn't even work...

The implementation is incomplete, as it stands, as it requires subclasses of UIResponder (such as UILabel or other UIView's) to implement tmlTranslationKeys method - which should return list of localizable strings.