Change translation deactivation behavior to show a list of options instead
Touch and hold with 4 fingers (on simulator - 2) to activate translation mode, but when the same gesture is used again, you'll get a list of options - Change Language, Deactivate Translation, and Cancel. This allow developer to simply add #import <TMLKit/TMLKit.h> inside the app delegate, and when responding to application launch, just call [TML sharedInstanceWithApplicationKey:accessToken:]. That is it!

At that point, turn on translation mode with aforementioned gesture - all found strings get uploaded to the API server. And you're on your way to change language and translate strings.

So far this only works with registered strings - strings from NIBs are automatically registered...