Make sure we get all translation keys, not just first page of multi-page result
Apparently some requests take "all=true" param and use that to return all the results, avoiding pagination - this seems to be working for translations, but not translation keys. So, when making a request with params that indicate to include all results, check if we get a paginated response, and paginate thru subsequent requests, collecting results into a running api response. Pagination info will be stripped from that running api response as it's irrelevant...