Some modifications on how we handle attributed strings:
* Skip all-encompassing decorated token on attributed strings - that's the case with all the plain strings set in IB - they come thru as attributed strings, with single attributed region spanning the whole string; that results in all strings like [style:...]
* When handling attributed strings - take into account strings that have no decorated tokens mentioned in the strings, but there might be a token with attributed styles - that would indicate it needs to be applied to the entire string

Originally this was a matter of convenience, but b/c IB does funny things when it comes to using strings in NIBs - this had to happen in order to deal with situations where strings are set on an object that are later applied to labels and show up as attributed, for example. And strings "[style:Title]" and "Title" will inevitably have different hash. That would result in odd/broken behavior when it comes to re-localizing things

* Also, objects that respond with localizable key paths - and there are two paths for the same property - one attributed and one plain - just check if we have attributed and use it if we have, defaulting to plain... No reason to parse TML decorated tokens or anything like that - it's just a waste of time