Reifying various methods of localizing strings, getting rid of as much magic as possible - giving users finer control over localization and translation.
No longer using keyPaths to capture identifiers for updating localized strings in real time, and getting rid of all methods that tried to automate the process - including walking the view hierarchy and finding objects, etc.

Objects that construct localized strings via provided macros register localized string and translation key in the internal tmlRegistry, and then register themselves with TML instance. This is done to accommodate in-app translation. The way that works is - when user taps on a screen, we find a view that contains localizable properties under that touch, we then find a string in those properties and match it to the appropriate translation key. TMLLocalizedString macros take care of capturing the sender object in the options that are passed on to the actual localization method, which takes care of registering required data.

For application to update localized strings in real-time - whether it's due to updated bundle files, or switching language, or making changes via inline translations - it needs to associate appropriate localized strings with reuse identifier. That will also capture whatever localization data was used to arrive at the current string, again, in the internal tmlRegistry object. Additional macros were added, ending with WithReuseIdentifier, making the reuse identifier argument explicit. Objects that create reusable strings via those macros automatically register themselves with TML instance, similarly to how normal localized string are handled, but in a separate table of objects. When localization data changes, and strings need to be updated, these objects are messaged updateTMLLocalizedStringWithInfo:forReuseIdentifier: passing all the required localization data, along with the new localization string. TMLReusableLocalization protocol was created to reify that method.

The default behavior of updateTMLLocalizedStringWithInfo:forReuseIdentifier: is to treat reuseIdentifier as a keyPath into the receiver - so as to accommodate automatic localization done with NIB objects.