Remove referencing to staging servers
Move those defines to TMLPrivateConfig scheme.
TMLPrivateConfig.h is now being generated by a build script. Git seems to reset the contents of the file as it's marked --assume-unchanged (so that the base file remains in the repo but no changes are being picked up by git).
To go around this issue, and still include a defined header file, and maintain defines privately - introducing a private plist file and a build script. You maintain TMLPrivateConfig.plist file with private defines and the build script will generate TMLPrivateConfig.h. The plist file isn't part of the repo and the build script will bail out if neither of the TMLPrivateConfig.* files are present